Init docker + refactoring
