+++ .env.example
@@ -1,2 +1,8 @@
 APP_HOST=127.0.0.1
 APP_PORT=8000
+
+DATABASE_DB=taskapi
+DATABASE_HOST=db
+DATABASE_USER=user123
+DATABASE_PASSWORD=password123
+DATABASE_PORT=5432
