+++ .gitignore
@@ -3,3 +3,4 @@ venv
 __pycache__
 /.env
 .idea
+postgres_data
