+++ docker-compose.yml
@@ -1,0 +1,26 @@
+version: "3.8"
+
+services:
+  taskapi:
+    build: /taskapi
+    container_name: taskapi
+    command: uvicorn app:app --host 0.0.0.0 --port 8000
+    ports:
+      - ${APP_PORT}:8000
+    depends_on:
+      - db
+    volumes:
+      - ./.env:/.env
+
+  db:
+    image: postgres
+    container_name: ${DATABASE_HOST}
+    environment:
+      - POSTGRES_USER=${DATABASE_USER}
+      - POSTGRES_PASSWORD=${DATABASE_PASSWORD}
+      - POSTGRES_DB=${DATABASE_DB}
+    command: postgres -p ${DATABASE_PORT}
+    ports:
+        - ${DATABASE_PORT}:${DATABASE_PORT}
+    volumes:
+      - ./postgres_data:/var/lib/postgresql/data
