+++ taskapi/Dockerfile
@@ -1,0 +1,10 @@
+FROM python:slim
+
+WORKDIR /taskapi
+
+COPY requirements.txt .
+RUN pip install -r requirements.txt
+
+COPY . .
+
+CMD ["uvicorn", "app:app", "--host", "0.0.0.0", "--port", "8000"]
